pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "infographic"
version = "0.1.0"
description = "CLI tool for generating AI-powered infographics for e-ink devices"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0",
    "google-genai>=1.0",
    "feedparser>=6.0",
    "Pillow>=11.0",
    "pydantic>=2.0",
    "httpx>=0.27",
]

[project.scripts]
infographic = "infographic.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/infographic"]
